Make code cleaner by moving structured data extraction into utils file
